nixos/nextcloud: drop httpd configuration from manual
Closes #243203

This hasn't changed in years because nobody maintains it. Considering
the amount of updates in the meantime to the nginx configuration, I'm
pretty sure that this doesn't work anymore.

I think it's OK to keep the part about how to really disable nginx, but
that's it.

The reason this also closes #243203 is that config in the manual saw
zero maintenance and I don't expect that to change. This should really
be maintained by interested people in a community wiki.